[wip] mobile welcome screen and project toolbrar component extraction
