+++ src/hooks/index.ts
@@ -13,6 +13,7 @@ export { useBreakpointObserver } from "./useBreakpointObserver";
 export { useMobileView } from "./useMobileView";
 export { useCategoryValidation } from "./useCategoryValidation";
 export { useNumberField } from "./useNumberField";
+export { useWindowSize } from "./useWindowSize";
 
 // useDefaultImage not included to encourage use of
 // import {useDefaultImage, DispatchLocation} from "hooks/useDefaultImage"
