+++ src/views/ProjectViewer/sections/ProjectAppBar/ProjectAppBar.tsx
@@ -1,60 +1,35 @@
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
 
+import { Box, Divider, Badge } from "@mui/material";
 import {
-  Slider,
-  Box,
-  Typography,
-  FormControl,
-  Menu,
-  Tooltip,
-  IconButton,
-  Chip,
-  Divider,
-  Badge,
-} from "@mui/material";
-import {
-  ZoomIn as ZoomInIcon,
-  Add as AddIcon,
-  Remove as RemoveIcon,
   Delete as DeleteIcon,
   Deselect as DeselectIcon,
-  Gesture as GestureIcon,
-  LabelOutlined as LabelOutlinedIcon,
   SelectAll as SelectAllIcon,
-  Straighten as StraightenIcon,
 } from "@mui/icons-material";
 
-import { useDialogHotkey, useHotkeys, useMenu, useMobileView } from "hooks";
+import { useDialogHotkey, useHotkeys, useMobileView } from "hooks";
 import { useThingSelection } from "../../hooks";
 
 import { LogoLoader, AlertBar } from "components/ui";
 import { TooltipTitle, TooltipButton } from "components/ui/tooltips";
 import { CustomAppBar } from "components/layout";
 import { ConfirmationDialog } from "components/dialogs/ConfirmationDialog";
-import { TextFieldWithBlur } from "components/inputs";
-import { ImageCategoryMenu } from "./ImageCategoryMenu";
+import { ZoomControl } from "./ZoomControl";
+import { ProjectTextField } from "./ProjextTextField";
+import { CategorizeChip } from "./CategorizeChip";
 
-import { projectSlice } from "store/project";
-import { applicationSettingsSlice } from "store/applicationSettings";
 import { dataSlice } from "store/data";
-import { selectActiveKindId, selectProjectName } from "store/project/selectors";
-import { selectActiveCategories } from "store/project/reselectors";
+import { selectActiveKindId } from "store/project/selectors";
 import {
   selectAlertState,
   selectLoadPercent,
-  selectLoadMessage,
 } from "store/applicationSettings/selectors";
 
 import { pluralize } from "utils/stringUtils";
-import { isUnknownCategory } from "store/data/utils";
 import { HotkeyContext } from "utils/enums";
-import { Partition } from "utils/models/enums";
-import { HelpItem } from "components/layout/HelpDrawer/HelpContent";
-
-const minZoom = 0.6;
-const maxZoom = 4;
+import { ImageViewerButton } from "./ImageViewerButton";
+import { MeasurementsButton } from "./MeasurementsButton";
 
 export const ProjectAppBar = () => {
   const dispatch = useDispatch();
@@ -62,7 +37,6 @@ export const ProjectAppBar = () => {
   const loadPercent = useSelector(selectLoadPercent);
   const alertState = useSelector(selectAlertState);
   const isMobile = useMobileView();
-  const navigate = useNavigate();
 
   const {
     allSelected,
@@ -85,18 +59,6 @@ export const ProjectAppBar = () => {
         disposeColorTensors: true,
       }),
     );
-  };
-
-  const handleNavigateImageViewer = () => {
-    navigate("/imageviewer", {
-      state: {
-        initialThingIds: allSelectedThingIds,
-      },
-    });
-  };
-
-  const handleNavigateMeasurements = () => {
-    navigate("/measurements");
   };
 
   useHotkeys(
@@ -188,36 +150,8 @@ export const ProjectAppBar = () => {
                 flexItem
                 sx={{ mr: 2 }}
               />
-              <Tooltip
-                title={
-                  allSelectedThingIds.length === 0
-                    ? "Select Objects to Annotate"
-                    : "Annotate Selection"
-                }
-              >
-                <span>
-                  <Chip
-                    data-help={HelpItem.NavigateImageViewer}
-                    avatar={<GestureIcon color="inherit" />}
-                    label="Annotate"
-                    onClick={handleNavigateImageViewer}
-                    variant="outlined"
-                    sx={{ marginRight: 1 }}
-                    disabled={allSelectedThingIds.length === 0}
-                  />
-                </span>
-              </Tooltip>
-              <Tooltip title="Go to Measurements">
-                <span>
-                  <Chip
-                    data-help={HelpItem.NavigateMeasurements}
-                    avatar={<StraightenIcon color="inherit" />}
-                    label="Measurements"
-                    onClick={handleNavigateMeasurements}
-                    variant="outlined"
-                  />
-                </span>
-              </Tooltip>
+              <ImageViewerButton selectedThings={allSelectedThingIds} />
+              <MeasurementsButton />
             </>
           )}
         </CustomAppBar>
@@ -237,187 +171,6 @@ export const ProjectAppBar = () => {
         isOpen={deleteImagesDialogisOpen}
         onClose={handleCloseDeleteImagesDialog}
       />
-    </>
-  );
-};
-
-const ZoomControl = () => {
-  const dispatch = useDispatch();
-  const [value, setValue] = useState<number>(1);
-  const { onOpen, onClose, open, anchorEl } = useMenu();
-
-  const handleSizeChange = (event: Event, newValue: number | number[]) => {
-    setValue(newValue as number);
-    dispatch(
-      applicationSettingsSlice.actions.updateTileSize({
-        newValue: newValue as number,
-      }),
-    );
-  };
-
-  const onZoomOut = () => {
-    const newValue = value - 0.1 >= minZoom ? value - 0.1 : minZoom;
-    setValue(newValue as number);
-    dispatch(
-      applicationSettingsSlice.actions.updateTileSize({
-        newValue: newValue as number,
-      }),
-    );
-  };
-
-  const onZoomIn = () => {
-    const newValue = value + 0.1 <= maxZoom ? value + 0.1 : maxZoom;
-    setValue(newValue as number);
-    dispatch(
-      applicationSettingsSlice.actions.updateTileSize({
-        newValue: newValue as number,
-      }),
-    );
-  };
-
-  return (
-    <>
-      <IconButton
-        data-help={HelpItem.GridZoom}
-        color="inherit"
-        onClick={onOpen}
-      >
-        <ZoomInIcon />
-      </IconButton>
-      <Menu anchorEl={anchorEl} open={open} onClose={onClose}>
-        <Box display="flex" flexDirection="column" alignItems="center">
-          <AddIcon onClick={onZoomIn} />
-          <Slider
-            orientation="vertical"
-            value={value}
-            min={minZoom}
-            max={maxZoom}
-            step={0.1}
-            onChange={handleSizeChange}
-            sx={{ height: (maxZoom - minZoom) * 20 + "px", my: 1, mr: 0 }}
-          />
-          <RemoveIcon onClick={onZoomOut} />
-        </Box>
-      </Menu>
-    </>
-  );
-};
-
-const CategorizeChip = ({
-  unfilteredSelectedThings,
-}: {
-  unfilteredSelectedThings: string[];
-}) => {
-  const dispatch = useDispatch();
-  const categories = useSelector(selectActiveCategories);
-  const [categoryMenuAnchorEl, setCategoryMenuAnchorEl] =
-    React.useState<null | HTMLElement>(null);
-  const onOpenCategoriesMenu = (event: React.MouseEvent<HTMLDivElement>) => {
-    setCategoryMenuAnchorEl(event.currentTarget);
-  };
-
-  const onCloseCategoryMenu = () => {
-    setCategoryMenuAnchorEl(null);
-  };
-  const handleUpdateCategories = (categoryId: string) => {
-    const updates = unfilteredSelectedThings.map((thingId) => ({
-      id: thingId,
-      categoryId: categoryId,
-      partition: isUnknownCategory(categoryId)
-        ? Partition.Inference
-        : Partition.Unassigned,
-    }));
-    dispatch(
-      dataSlice.actions.updateThings({
-        updates,
-      }),
-    );
-  };
-  return (
-    <>
-      <Tooltip
-        title={
-          unfilteredSelectedThings.length === 0
-            ? "Select Objects to Categorize"
-            : "Categorize Selection"
-        }
-      >
-        <span>
-          <Chip
-            data-help={HelpItem.Categorize}
-            avatar={<LabelOutlinedIcon color="inherit" />}
-            label="Categorize"
-            onClick={onOpenCategoriesMenu}
-            variant="outlined"
-            sx={{ marginRight: 1 }}
-            disabled={unfilteredSelectedThings.length === 0}
-          />
-        </span>
-      </Tooltip>
-      <ImageCategoryMenu
-        anchorEl={categoryMenuAnchorEl as HTMLElement}
-        selectedIds={unfilteredSelectedThings}
-        onClose={onCloseCategoryMenu}
-        open={Boolean(categoryMenuAnchorEl as HTMLElement)}
-        onUpdateCategories={handleUpdateCategories}
-        categories={categories}
-      />
-    </>
-  );
-};
-
-const ProjectTextField = () => {
-  const dispatch = useDispatch();
-
-  const loadMessage = useSelector(selectLoadMessage);
-  const projectName = useSelector(selectProjectName);
-  const [newProjectName, setNewProjectName] = useState<string>(projectName);
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const handleTextFieldBlur = () => {
-    if (projectName === newProjectName) return;
-    dispatch(projectSlice.actions.setProjectName({ name: newProjectName }));
-    setNewProjectName("");
-  };
-
-  const handleTextFieldChange = (
-    event: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    setNewProjectName(event.target.value);
-  };
-
-  useEffect(() => {
-    setNewProjectName(projectName);
-  }, [projectName]);
-
-  return (
-    <>
-      {loadMessage ? (
-        <Typography ml={5} sx={{ flexGrow: 1 }}>
-          {loadMessage}
-        </Typography>
-      ) : (
-        <FormControl>
-          <TextFieldWithBlur
-            data-help={HelpItem.ProjectName}
-            onChange={handleTextFieldChange}
-            onBlur={handleTextFieldBlur}
-            value={newProjectName}
-            inputRef={inputRef}
-            size="small"
-            sx={{ ml: 5 }}
-            variant="standard"
-            slotProps={{
-              htmlInput: { min: 0 },
-              input: {
-                slotProps: {
-                  input: { min: 0 },
-                },
-              },
-            }}
-          />
-        </FormControl>
-      )}
     </>
   );
 };
